web/requirements.txt
pandas==3.0.2
Pillow==12.2.0
python-dotenv==1.2.2
streamlit==1.56.0
matplotlib==3.10.8
ipython==9.10.0
prometheus-client==0.21.1